BubbleNotifyConfigPanel: -added gui settings for BUBBLE_NOTIFY_ON_NEW_LINKGRABBER_LINKS_END_NOTIFY_DELAY -fixes thread 82574
git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@43357 ebf7c1c2-ba36-0410-9fe8-c592906822b4

+++ src/org/jdownloader/gui/notify/linkcrawler/LinkCrawlerBubbleContent.java
@@ -35,7 +35,7 @@ public class LinkCrawlerBubbleContent extends AbstractBubbleContentPanel {
     private Pair             listQueue;
     private Pair             packages;
     private Pair             online;
-    private final long       CLOSETIMEOUT = JsonConfig.create(BubbleNotifyConfig.class).getBubbleNotifyOnNewLinkgrabberLinksEndNotifyDelay();
+    private final int        CLOSETIMEOUT = JsonConfig.create(BubbleNotifyConfig.class).getBubbleNotifyOnNewLinkgrabberLinksEndNotifyDelay();
     private final LinkOrigin origin;
 
     public LinkCrawlerBubbleContent(JobLinkCrawler crawler) {
